Creates the basic searching algorithm
WARNING: search_grid.js still needs to export the search() function
in order to be usable in other files.

The search() function takes in 2 mandatory arguments and 2
optional arguments (respectively):

- the grid of letters (mandatory)
- the word to be searched (mandatory)
- the last word that has been searched (opt)
- the result from the last word searched (opt)

It will also return an output in the form of a list of objects
containing the starting x and y coordinates of the word, as well as
the direction of the word and it's length.

If this output, along with the word associated with it, are included in
the input of a subsequent search that uses the same grid and an
an extension of the previous word, a quicker process will be used to
find the results.