Adds eslint and some basic rules
Please run the following commands in the root directory of this
project order to install/update necessary dependencies, including
.eslint:
npm install
npm upgrade

This commit adds eslint to package, package-lock and the config
file .eslintrc. For more documentation on how to configure .eslint
go to their website: https://eslint.org/docs/rules/

The goal is that in the future all code follows the same standards
and conventions.